requirements.txt
aws-cdk-lib>=2.104.0
constructs>=10.02.69
aws-cdk.aws-lambda-python-alpha>=2.51.0a0
cdk-nag>=2.27.88
